does a lot regarding the bar remote

+++ code/game/objects/items/miscellaneous.dm
@@ -136,32 +136,69 @@
 	name = "bar remote"
 	desc = "A fancy-looking remote that instructs the bar's nanite systems to construct a layout."
 	uses = 1
+	var/unusable = FALSE
+
+/obj/item/choice_beacon/bar_remote/Initialize()
+	. = ..()
+	addtimer(CALLBACK(src, .proc/trigger_random), 3000) // If no one makes an option by the time we are 5 minutes into the round, choose random
+
+/obj/item/choice_beacon/bar_remote/proc/trigger_random()
+	if(SSevents.holidays && SSevents.holidays["St. Patrick's Day"])
+		spawn_option("Bar Irish")
+	else
+		spawn_option("Random")
+	unusable = TRUE
+	world.log << "BAR DONE"
+	world.log << "BAR DONE"
+	world.log << "BAR DONE"
+	world.log << "BAR DONE"
+	world.log << "BAR DONE"
+	world.log << "BAR DONE"
+	world.log << "BAR DONE"
+	world.log << "BAR DONE"
 
 /obj/item/choice_beacon/bar_remote/generate_display_names()
 	var/static/list/bar_list
 	if(!bar_list)
 		bar_list = list()
-		var/list/templist = list("Bar Trek", "Bar Spacious", "Bar Box", "Bar Casino", "Bar Citadel", "Bar Conveyor", "Bar Diner", "Bar Disco", "Bar Purple", "Bar Cheese", "Bar Clock", "Bar Arcade")
+		var/list/templist
+		if(SSevents.holidays && SSevents.holidays["St. Patrick's Day"])
+			templist += "Bar Irish"
+		else
+			templist = GLOB.potential_bars
+			templist += "Random"
 		for(var/V in templist)
 			bar_list[V] = V
 	return bar_list
 
-/obj/item/choice_beacon/bar_remote/spawn_option(choice,mob/living/M)
-	to_chat(M, "Bar choice locked in. Construction nanites engaged.")
+/obj/item/choice_beacon/bar_remote/spawn_option(choice, mob/living/M)
+	if(unusable)
+		return
+
+	if(!choice)
+		return
+	
+	if(M)
+		to_chat(M, "<span class='notice'>Bar choice locked in. Construction nanites engaged.</span>")
+	
+	if(choice == "Random")
+		choice = pick(GLOB.potential_bars)
+
 	var/datum/map_template/template = SSmapping.station_room_templates[choice]
 
-	for(var/obj/effect/landmark/stationroom/box/bar/B in GLOB.landmarks_list )
+	for(var/obj/effect/landmark/stationroom/box/bar/B in GLOB.landmarks_list)
 		template.load(B.loc, centered = FALSE)
 		qdel(B)
 	
 	// Reboots lighting because it breaks on load
-	var/area/K = GLOB.areas_by_type[/area/crew_quarters/kitchen]
-	K.set_dynamic_lighting(DYNAMIC_LIGHTING_DISABLED)
-	K.set_dynamic_lighting()
+	for(var/area/crew_quarters/kitchen/K in world)
+		K.set_dynamic_lighting(DYNAMIC_LIGHTING_DISABLED)
+		K.set_dynamic_lighting()
 	
 	var/area/B = GLOB.areas_by_type[/area/crew_quarters/bar]
 	B.set_dynamic_lighting(DYNAMIC_LIGHTING_DISABLED)
 	B.set_dynamic_lighting()
+	qdel(src)
 
 /obj/item/skub
 	desc = "It's skub."
